Fix base amount styles
